requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.2.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
Mako==1.3.9
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
passlib==1.7.4
pydantic==2.10.6
pydantic-core==2.27.2
PyJWT==2.9.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.44.0
typing-extensions==4.12.2
uvicorn==0.33.0
zipp==3.20.2
